overflow works

+++ BrainfuckHeader.java
@@ -1,8 +1,8 @@
 import java.io.IOException;
 
 public class %s {
-    private static byte[] data = new byte[256];
-    private static byte ptr = 0;
+    private static short[] data = new short[256];
+    private static short ptr = 0;
     
     public static void main(String[] args) throws IOException {
         // end of header
